fix: publish shared capa fixtures from a testing package
Downstream test suites import capa fixtures from this distribution, but no
1.0.x wheel contains them: `[tool.setuptools.packages.find]`'s `*tests*`
exclusion drops `xblocks_contrib.problem.capa.tests` entirely. MANIFEST.in
tries to hold them back --

    # We do this because openedx-platform imports these specific test helpers.
    include xblocks_contrib/problem/capa/tests/test_util.py
    ...

-- but MANIFEST.in only governs the sdist, so the intent never reached the
wheel. openedx-platform consequently fails to collect ~17 of its own test
modules against any 1.0.x release, and stays pinned to 0.17.0.

Move the two fixtures consumers actually import into a `capa/testing/`
package. It is deliberately not under a `tests/` directory, so the existing
exclusion cannot match it -- no packaging configuration changes at all:

* `capa/testing/response_xml_factory.py` -- moved verbatim; a pure factory
  module with no tests in it.
* `capa/testing/codejail.py` -- `UseUnsafeCodejail`, lifted out of
  `tests/test_util.py`. That module mixes a real test case (`UtilTest`) with a
  fixture only consumers use, so it is split rather than moved; `UtilTest`
  stays behind.

`capa/tests/helpers.py` stays internal: despite MANIFEST.in listing it, nothing
outside this repo imports it, and its `load_fixture` resolves paths relative to
`tests/test_files`. The now-stale MANIFEST.in block is removed -- one of the
three files it names no longer exists, and none of them need re-including.

Consumers change import paths as follows:

    from xblocks_contrib.problem.capa.tests.response_xml_factory import X
    -> from xblocks_contrib.problem.capa.testing.response_xml_factory import X

    from xblocks_contrib.problem.capa.tests.test_util import UseUnsafeCodejail
    -> from xblocks_contrib.problem.capa.testing.codejail import UseUnsafeCodejail

No released 1.0.x wheel exposes the old paths, so nothing installable can break.
