+++ xblocks_contrib/problem/capa/testing/__init__.py
@@ -1,0 +1,7 @@
+"""
+Reusable capa testing utilities.
+
+Unlike ``xblocks_contrib.problem.capa.tests``, this package ships in the
+published distribution: it holds the fixtures that downstream test suites --
+notably openedx-platform's -- import. Keep actual tests out of it.
+"""
